+++ includes/class-bitbucket-api.php
@@ -27,10 +27,6 @@ class GitHub_Updater_BitBucket_API extends GitHub_Updater {
 		$this->type  = $type;
 		self::$hours = 4;
 
-		if ( ! empty( $this->type->timeout ) ) {
-			self::$hours = (float) $this->type->timeout;
-		}
-
 		add_filter( 'http_request_args', array( $this, 'maybe_authenticate_http' ), 10, 2 );
 	}
 
@@ -47,10 +43,10 @@ class GitHub_Updater_BitBucket_API extends GitHub_Updater {
 	 */
 	protected function api( $url ) {
 		$response = wp_remote_get( $this->get_api_url( $url ) );
+		$code     = wp_remote_retrieve_response_code( $response );
 
-		if ( is_wp_error( $response ) || ( ( '200' || '404' ) != wp_remote_retrieve_response_code( $response ) ) ) {
-			return false;
-		}
+		if ( is_wp_error( $response ) ) { return false; }
+		if ( ( 200 || 404 )!= $code ) { return false; }
 
 		return json_decode( wp_remote_retrieve_body( $response ) );
 	}
@@ -106,13 +102,13 @@ class GitHub_Updater_BitBucket_API extends GitHub_Updater {
 	 * @since 1.0.0
 	 */
 	public function get_remote_info( $file ) {
-		$response = get_site_transient( 'ghu-' . md5( $this->type->repo . $file ) );
+		$response = $this->get_transient( $file );
 
 		if ( ! $response && isset( $this->type->branch ) ) {
 			$response = $this->api( '1.0/repositories/:owner/:repo/src/' . trailingslashit($this->type->branch) . $file );
 
 			if ( $response ) {
-				set_site_transient( 'ghu-' . md5( $this->type->repo . $file ), $response, ( self::$hours * HOUR_IN_SECONDS ) );
+				$this->make_transient( $file, $response );
 			}
 		}
 
@@ -163,7 +159,7 @@ class GitHub_Updater_BitBucket_API extends GitHub_Updater {
 	 */
 	public function get_remote_tag() {
 		$download_link_base = 'https://bitbucket.org/' . trailingslashit( $this->type->owner ) . $this->type->repo . '/get/';
-		$response           = get_site_transient( 'ghu-' . md5( $this->type->repo . 'tags' ) );
+		$response           = $this->get_transient( 'tags' );
 
 		if ( ! $response ) {
 			$response = $this->api( '1.0/repositories/:owner/:repo/tags' );
@@ -174,7 +170,7 @@ class GitHub_Updater_BitBucket_API extends GitHub_Updater {
 			}
 
 			if ( $response ) {
-				set_site_transient( 'ghu-' . md5( $this->type->repo . 'tags' ), $response, ( self::$hours * HOUR_IN_SECONDS ) );
+				$this->make_transient( 'tags', $response );
 			}
 		}
 
@@ -247,7 +243,7 @@ class GitHub_Updater_BitBucket_API extends GitHub_Updater {
 			require_once 'markdown.php';
 		}
 
-		$response = get_site_transient( 'ghu-' . md5( $this->type->repo . 'changes' ) );
+		$response = $this->get_transient( 'changes' );
 
 		if ( ! $response ) {
 			$response = $this->api( '1.0/repositories/:owner/:repo/src/' . trailingslashit($this->type->branch) . $changes  );
@@ -258,15 +254,14 @@ class GitHub_Updater_BitBucket_API extends GitHub_Updater {
 			}
 
 			if ( $response ) {
-				set_site_transient( 'ghu-' . md5( $this->type->repo . 'changes' ), $response, ( self::$hours * HOUR_IN_SECONDS ) );
+				$this->make_transient( 'changes', $response );
 			}
 		}
 
 		if ( ! $response ) { return false; }
 		if ( isset( $response->message ) ) { return false; }
 
-		$changelog = '';
-		$changelog = get_site_transient( 'ghu-' . md5( $this->type->repo . 'changelog' ), $changelog );
+		$changelog = $this->get_transient( 'changelog' );
 
 		if ( ! $changelog ) {
 			if ( function_exists( 'Markdown' ) ) {
@@ -274,7 +269,7 @@ class GitHub_Updater_BitBucket_API extends GitHub_Updater {
 			} else {
 				$changelog = '<pre>' . $response->data . '</pre>';
 			}
-			set_site_transient( 'ghu-' . md5( $this->type->repo . 'changelog' ), $changelog, ( self::$hours * HOUR_IN_SECONDS ) );
+			$this->make_transient( 'changelog', $changelog );
 		}
 
 		$this->type->sections['changelog'] = $changelog;
@@ -288,13 +283,13 @@ class GitHub_Updater_BitBucket_API extends GitHub_Updater {
 	 * @return base64 decoded repository meta data
 	 */
 	public function get_repo_meta() {
-		$response = get_site_transient( 'ghu-' . md5( $this->type->repo . 'meta' ) );
+		$response = $this->get_transient( 'meta' );
 
 		if ( ! $response ) {
 			$response = $this->api( '2.0/repositories/:owner/:repo' );
 
 			if ( $response ) {
-				set_site_transient( 'ghu-' . md5( $this->type->repo . 'meta' ), $response, ( self::$hours * HOUR_IN_SECONDS ) );
+				$this->make_transient( 'meta', $response );
 			}
 		}